+++ src/main/java/com/almostreliable/summoningrituals/compat/viewer/jei/entity/EntityIngredient.java
@@ -1,5 +1,6 @@
 package com.almostreliable.summoningrituals.compat.viewer.jei.entity;
 
+import com.almostreliable.summoningrituals.compat.kubejs.builder.SummoningEntityBuilder;
 import com.almostreliable.summoningrituals.recipe.EntityInfo;
 
 import net.minecraft.client.Minecraft;
@@ -11,13 +12,15 @@ import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.SpawnEggItem;
 
 import com.mojang.serialization.Codec;
+import com.mojang.serialization.codecs.RecordCodecBuilder;
 
 import org.jetbrains.annotations.Nullable;
 
 public final class EntityIngredient {
 
-    // TODO: improve codec to only encode entity type holder
-    public static final Codec<EntityIngredient> CODEC = EntityInfo.CODEC.xmap(EntityIngredient::new, EntityIngredient::getEntityInfo);
+    public static final Codec<EntityIngredient> CODEC = RecordCodecBuilder.create(i -> i.group(
+        BuiltInRegistries.ENTITY_TYPE.holderByNameCodec().fieldOf(Entity.ID_TAG).forGetter(e -> e.entityInfo.entity())
+    ).apply(i, e -> new EntityIngredient(new SummoningEntityBuilder(e).build())));
 
     private final EntityInfo entityInfo;
     @Nullable