+++ src/main/java/com/almostreliable/summoningrituals/compat/viewer/jei/entity/EntityIngredientRenderer.java
@@ -1,6 +1,8 @@
 package com.almostreliable.summoningrituals.compat.viewer.jei.entity;
 
+import com.almostreliable.summoningrituals.SummoningRituals;
 import com.almostreliable.summoningrituals.client.MeasuringBufferSource;
+import com.almostreliable.summoningrituals.client.MeasuringBufferSource.MeasuringResult;
 
 import net.minecraft.ChatFormatting;
 import net.minecraft.client.Minecraft;
@@ -8,6 +10,7 @@ import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.client.renderer.LightTexture;
 import net.minecraft.network.chat.Component;
 import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.entity.boss.wither.WitherBoss;
 import net.minecraft.world.item.TooltipFlag;
 
 import com.mojang.blaze3d.platform.Lighting;
@@ -15,17 +18,32 @@ import com.mojang.blaze3d.systems.RenderSystem;
 import com.mojang.blaze3d.vertex.PoseStack;
 import com.mojang.math.Axis;
 import mezz.jei.api.ingredients.IIngredientRenderer;
+import org.joml.Vector3f;
 
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-public class EntityIngredientRenderer implements IIngredientRenderer<EntityIngredient> {
+public final class EntityIngredientRenderer implements IIngredientRenderer<EntityIngredient> {
 
+    public static final EntityIngredientRenderer BOOKMARK_RENDERER = new EntityIngredientRenderer(true, false);
+    public static final EntityIngredientRenderer INPUT_RENDERER = new EntityIngredientRenderer(false, true);
+    public static final EntityIngredientRenderer OUTPUT_RENDERER = new EntityIngredientRenderer(true, true);
     private static final int TEXT_COLOR = 16_777_215;
+    private static final int MEASURE_TICKS = 40;
+    private static final int HALF_ROT = 180;
+    private static final int SLOT_SIZE = 16;
+
     private final Minecraft mc = Minecraft.getInstance();
-    private final Map<String, MeasuringResult> measurementCache = new HashMap<>();
+    private final Map<String, MeasuringResult> measuringResultCache = new HashMap<>();
+    private final boolean scissor;
+    private final boolean renderCount;
+
+    private EntityIngredientRenderer(boolean scissor, boolean renderCount) {
+        this.scissor = scissor;
+        this.renderCount = renderCount;
+    }
 
     @Override
     public void render(GuiGraphics guiGraphics, EntityIngredient entityIngredient) {
@@ -33,14 +51,19 @@ public class EntityIngredientRenderer implements IIngredientRenderer<EntityIngre
             return;
         }
 
+        var entityId = entityIngredient.getResourceLocation().toString();
+        var measuringResult = measureEntity(entity, entityId);
+        if (measuringResult == MeasuringResult.EMPTY) return;
+
         entity.tickCount = mc.player.tickCount;
 
         var poseStack = guiGraphics.pose();
         poseStack.pushPose();
         {
-            renderEntity(guiGraphics, entityIngredient, entity, poseStack);
+            renderEntity(guiGraphics, poseStack, entity, measuringResult);
         }
         poseStack.popPose();
+        if (!renderCount) return;
         poseStack.pushPose();
         {
             renderCount(guiGraphics, entityIngredient, poseStack);
@@ -49,41 +72,80 @@ public class EntityIngredientRenderer implements IIngredientRenderer<EntityIngre
     }
 
     @SuppressWarnings("deprecation")
-    private void renderEntity(
-        GuiGraphics guiGraphics, EntityIngredient entityIngredient, LivingEntity entity, PoseStack poseStack
-    ) {
-        var entityId = entityIngredient.getResourceLocation().toString();
+    private MeasuringResult measureEntity(LivingEntity entity, String entityId) {
+        var cached = measuringResultCache.get(entityId);
+        if (cached != null) return cached;
 
-        poseStack.translate(8, 16, 100);
-        poseStack.mulPose(Axis.ZP.rotationDegrees(180));
+        var entityRenderer = mc.getEntityRenderDispatcher();
+        var measuringBuffer = new MeasuringBufferSource();
+        var poseStack = new PoseStack();
+
+        // measure for 40 render ticks because size can change with animation
+        for (var i = 0; i < MEASURE_TICKS; i++) {
+            var ticks = i;
+            RenderSystem.runAsFancy(() -> entityRenderer.render(
+                entity, 0, 0, 0, 0, ticks, poseStack, measuringBuffer, LightTexture.FULL_BRIGHT
+            ));
+        }
+
+        var measuringResult = measuringBuffer.getData();
+        if (measuringResult == null) {
+            SummoningRituals.LOGGER.error("failed to measure entity: {}", entityId);
+            return MeasuringResult.EMPTY;
+        }
+        measuringResultCache.put(entityId, measuringResult);
+        return measuringResult;
+    }
+
+    @SuppressWarnings("deprecation")
+    private void renderEntity(
+        GuiGraphics guiGraphics, PoseStack poseStack, LivingEntity entity, MeasuringResult measuringResult
+    ) {
+        if (scissor) {
+            // gui graphics scissor doesn't take pose stack into account, bruh
+            var absolutePos = poseStack.last().pose().transformPosition(0, 0, 0, new Vector3f());
+            var absX = (int) absolutePos.x;
+            var absY = (int) absolutePos.y;
+            guiGraphics.enableScissor(absX, absY, absX + SLOT_SIZE, absY + SLOT_SIZE);
+        }
+
+        poseStack.translate(SLOT_SIZE / 2f, SLOT_SIZE, 100);
+        poseStack.mulPose(Axis.ZP.rotationDegrees(HALF_ROT));
 
         Lighting.setupForEntityInInventory();
         var entityRenderer = mc.getEntityRenderDispatcher();
         entityRenderer.setRenderShadow(false);
         RenderSystem.enableBlend();
 
-        var measuringResult = measurementCache.computeIfAbsent(
-            entityId, id -> {
-                var measuringBuffer = new MeasuringBufferSource();
-                RenderSystem.runAsFancy(() -> entityRenderer.render(
-                    entity, 0, 0, 0, 0, 1, poseStack, measuringBuffer,
-                    LightTexture.FULL_BRIGHT
-                ));
-
-                if (measuringBuffer.hasData()) {
-                    return new MeasuringResult(measuringBuffer.diagonal(), measuringBuffer.height());
-                }
-
-                var size = (float) entity.getBoundingBox().getSize();
-                return new MeasuringResult(size, 0);
+        // need to turn all mobs to the other side because the dragon and the bat have reversed models
+        // this requires inverting the models later by a negative z-value;
+        // tried so many other approaches, this is the only thing that works
+        entity.absRotateTo(HALF_ROT, 0);
+        entity.yBodyRotO = HALF_ROT;
+        entity.yBodyRot = HALF_ROT;
+        entity.yHeadRotO = HALF_ROT;
+        entity.yHeadRot = HALF_ROT;
+        if (entity instanceof WitherBoss witherBoss) {
+            // this is usually handled in the wither AI for whatever reason, Mojank
+            var yRotHeads = witherBoss.yRotHeads;
+            var yRotOHeads = witherBoss.yRotOHeads;
+            for (var i = 0; i < yRotHeads.length; i++) {
+                yRotHeads[i] = HALF_ROT;
+                yRotOHeads[i] = HALF_ROT;
             }
-        );
+        }
 
-        poseStack.translate(0, measuringResult.height(), 0);
+        var width = measuringResult.maxX() - measuringResult.minX();
+        var height = measuringResult.maxY() - measuringResult.minY();
+
+        var heightScale = SLOT_SIZE / height;
+        var widthScale = (SLOT_SIZE / width) * 2;
+        var scale = Math.min(widthScale, heightScale);
+
+        poseStack.translate(0, -measuringResult.minY() * heightScale, 0);
         poseStack.mulPose(Axis.YP.rotationDegrees(20));
         poseStack.mulPose(Axis.XP.rotationDegrees(5));
-        float scale = 18f / (float) Math.pow(measuringResult.diagonal(), 0.9);
-        poseStack.summoning$scale(scale);
+        poseStack.scale(scale, scale, -scale);
 
         RenderSystem.runAsFancy(() -> entityRenderer.render(
             entity, 0, 0, 0, 0, 1, poseStack, guiGraphics.bufferSource(),
@@ -93,6 +155,7 @@ public class EntityIngredientRenderer implements IIngredientRenderer<EntityIngre
 
         entityRenderer.setRenderShadow(true);
         Lighting.setupFor3DItems();
+        if (scissor) guiGraphics.disableScissor();
     }
 
     private void renderCount(GuiGraphics guiGraphics, EntityIngredient entityIngredient, PoseStack poseStack) {
@@ -118,6 +181,4 @@ public class EntityIngredientRenderer implements IIngredientRenderer<EntityIngre
         }
         return tooltip;
     }
-
-    private record MeasuringResult(float diagonal, float height) {}
 }